fix:folder structure and best pratice code
