+++ app/Modules/Auth/Request/LoginUserRequest.php
@@ -2,6 +2,7 @@
 
 namespace App\Modules\Auth\Request;
 
+use App\Modules\Auth\DTOs\LoginUserDto;
 use App\Modules\Auth\Rules\ValidUserCredential;
 use Illuminate\Foundation\Http\FormRequest;
 
@@ -13,5 +14,10 @@ class LoginUserRequest extends FormRequest
             'email' => ['required', 'email', 'exists:users,email'],
             'password' => ['required', new ValidUserCredential],
         ];
+    }
+
+    public function validatedLogin()
+    {
+        return LoginUserDto::fromRequest($this->validated());
     }
 }
