+++ routes/api/auth/v1.php
@@ -1,0 +1,11 @@
+<?php
+
+use App\Http\Auth\Controllers\AuthController;
+use Illuminate\Support\Facades\Route;
+
+Route::post('/auth/register', [AuthController::class, 'register']);
+Route::post('/auth/login', [AuthController::class, 'login']);
+
+Route::middleware('auth:sanctum')->group(function () {
+    Route::delete('/auth/logout', [AuthController::class, 'logout']);
+});
